Conflicts:
	README.md
	build.xml
	src/protocolsupport/injector/network/NettyInjector.java
	src/protocolsupport/protocol/PacketDataSerializer.java
	src/protocolsupport/protocol/typeremapper/id/IdRemapper.java
	src/protocolsupport/protocol/typeremapper/watchedentity/remapper/SpecificType.java
	src/protocolsupport/utils/DataWatcherObject.java
	src/protocolsupport/utils/Utils.java

Update to 1.9, 1.9 items are not remapped yet, so crafting doesn't work.

+++ src/protocolsupport/protocol/transformer/middlepacket/clientbound/play/MiddleEntityRelMove.java
@@ -14,9 +14,9 @@ public abstract class MiddleEntityRelMove<T> extends MiddleEntity<T> {
 	@Override
 	public void readFromServerData(PacketDataSerializer serializer) throws IOException {
 		super.readFromServerData(serializer);
-		relX = serializer.readByte();
-		relY = serializer.readByte();
-		relZ = serializer.readByte();
+		relX = serializer.readShort();
+		relY = serializer.readShort();
+		relZ = serializer.readShort();
 		onGround = serializer.readBoolean();
 	}
