+++ src/protocolsupport/protocol/transformer/middlepacket/clientbound/play/MiddleSpawnGlobal.java
@@ -7,17 +7,17 @@ public abstract class MiddleSpawnGlobal<T> extends ClientBoundMiddlePacket<T> {
 
 	protected int entityId;
 	protected int type;
-	protected int x;
-	protected int y;
-	protected int z;
+	protected double x;
+	protected double y;
+	protected double z;
 
 	@Override
 	public void readFromServerData(PacketDataSerializer serializer) {
 		entityId = serializer.readVarInt();
 		type = serializer.readUnsignedByte();
-		x = serializer.readInt();
-		y = serializer.readInt();
-		z = serializer.readInt();
+		x = serializer.readDouble();
+		y = serializer.readDouble();
+		z = serializer.readDouble();
 	}
 
 }
